prepare android for shell: extract common methods and create renderer interface
