+++ src/SharedTransitions/Platforms/Android/Renderers/ITransitionRenderer.cs
@@ -1,0 +1,16 @@
+﻿using Android.Support.V4.App;
+using Xamarin.Forms;
+
+namespace Plugin.SharedTransitions.Platforms.Android
+{
+	public interface ITransitionRenderer
+	{
+		FragmentManager FragmentManager { get; set; }
+		string SelectedGroup { get; set; }
+		BackgroundAnimation BackgroundAnimation { get; set; }
+		Page PropertiesContainer { get; set; }
+		Page LastPageInStack { get; set; }
+		ITransitionMapper TransitionMap { get; set; }
+		global::Android.Transitions.Transition InflateTransitionInContext();
+	}
+}
